refactor(gui): extract launch runtime module
